+++ react/hooks/useNewtailMediaSearch/formatRequestBody.tsx
@@ -1,45 +1,110 @@
 import { useMemo } from 'react'
 
 import { useDeviceType } from '../useDeviceType'
-import { useQueryTerm } from '../useQueryTerm'
 import { useSessionData } from '../useSessionData'
+import { usePageContext } from '../usePageContext'
 
 type FormatRequestBodyProps = {
-  skus: string[]
+  skus?: string[]
   placement: string
+  mediaSize?: string
   quantityAds?: number
+  adType?: AdTypes
 }
 
 type FormatRequestBody = (props: FormatRequestBodyProps) => null | RequestBody
+
+type ContextData = {
+  category: CategoryContextData
+  brand_page: BrandPageContextData
+  search: SearchContextData
+  home: HomeContextData
+  product_page: ProductPageContextData
+}
+
+const adTypesKeys = {
+  product: 'product',
+  banner: 'banner',
+} as { [key in AdTypes]: AdTypes }
 
 export const useRequestBody: FormatRequestBody = ({
   skus,
   placement = 'search',
   quantityAds = 20,
+  mediaSize,
+  adType = 'product',
 }) => {
   const { sessionId, userId } = useSessionData()
-  const term = useQueryTerm()
+  const { context, term, brandName, categoryName } = usePageContext()
   const device = useDeviceType()
 
-  return useMemo(
+  const contextData: ContextData = useMemo(
+    () => ({
+      category: {
+        context: 'category',
+        category_name: categoryName || '',
+      },
+      brand_page: {
+        context: 'brand_page',
+        brand_name: brandName || '',
+      },
+      search: {
+        context: 'search',
+        term,
+        skus,
+      },
+      home: {
+        context: 'home',
+      },
+      product_page: {
+        context: 'product_page',
+        product_sku: '',
+      },
+    }),
+    [brandName, categoryName, skus, term]
+  )
+
+  const placementData = useMemo(
+    () => ({
+      [adTypesKeys.banner]: {
+        size: mediaSize,
+        quantity: quantityAds,
+        types: [adTypesKeys.banner],
+      },
+      [adTypesKeys.product]: {
+        quantity: quantityAds,
+        types: [adTypesKeys.product],
+      },
+    }),
+    [mediaSize, quantityAds]
+  )
+
+  const body = useMemo(
     () =>
       !sessionId
         ? null
         : ({
-            skus,
             device,
             term,
             user_id: userId,
             session_id: sessionId,
-            context: 'search',
             placements: {
-              [placement]: {
-                size: device,
-                quantity: quantityAds,
-                types: ['product'],
-              },
+              [placement]: placementData[adType],
             },
+            ...contextData?.[context],
           } as RequestBody),
-    [device, placement, quantityAds, sessionId, skus, term, userId]
+    [
+      adType,
+      context,
+      contextData,
+      device,
+      placement,
+      placementData,
+      sessionId,
+      term,
+      userId,
+    ]
   )
+
+  return body
 }